// src/services/api/generated/model/recipeIngredientResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Stock API
 * API de gestion de stock (produits, catégories, recettes, utilisateurs)
 * OpenAPI spec version: v0.0.1
 */
import type { RecipeProductReferenceResponse } from './recipeProductReferenceResponse';
import type { RecipeUnitReferenceResponse } from './recipeUnitReferenceResponse';

export interface RecipeIngredientResponse {
  product?: RecipeProductReferenceResponse;
  quantity?: number;
  unit?: RecipeUnitReferenceResponse;
}